ShoppingCartCommandHandlerIntegrationTests - fixed all tests and added idempotent tests; refactored ProcessManager (support for PM restore by causationId); code cleanup.

+++ src/Modules/Sales/Sales.Domain/Models/ShoppingCart/ShoppingCartItem.cs
@@ -77,7 +77,7 @@ namespace VShop.Modules.Sales.Domain.Models.ShoppingCart
         
         public static class Settings
         {
-            public const decimal MaxQuantityPerProduct = 10;
+            public const int MaxQuantityPerProduct = 10;
         }
     }
 }
